[Hotfix/Interim] Неправильное закрытие по Esc
