+++ src/components/app/app.tsx
@@ -65,12 +65,22 @@ const App : FC = () => {
   const isPublicFeedOpen = useSelector((state) => state.feed.public.isOpen);
   const isPrivateFeedOpen = useSelector((state) => state.feed.private.isOpen);
   const { errorMessage, successMessage } = useSelector((state) => state.api);
+
+  useEffect(() => {
+    console.log('Location has changed!');
+    console.dir(location);
+  }, [location]);
+
   const handleIngredientDetailsClose = () => {
+    console.log('Closing ings modal...');
+    console.dir(location);
     dispatch(releaseIngredient());
     history.push({
       ...location.state.background as TLocationState | TLocation,
       state: { background: null },
     });
+    console.log('history pushed!');
+    console.dir(location);
   };
   const orderDetailsClose = () => {
     dispatch(releaseOrder());
