+++ src/components/modal-overlay/modal-overlay.tsx
@@ -1,4 +1,4 @@
-import React, { FC, MouseEventHandler } from 'react';
+import React, { FC, MouseEventHandler, useEffect } from 'react';
 
 import moStyles from './modal-overlay.module.css';
 import { TModalOverlayProps } from '../../types/components.props.types';
@@ -9,6 +9,17 @@ const ModalOverlay : FC<TModalOverlayProps> = ({ onClose }) => {
       onClose();
     }
   };
+  useEffect(() => {
+    const handleEscClose = (evt : KeyboardEvent) => {
+      if (evt.key === 'Escape') {
+        onClose();
+      }
+    };
+    document.addEventListener('keydown', handleEscClose);
+    return () => {
+      document.removeEventListener('keydown', handleEscClose);
+    };
+  }, [onClose]);
   return (
     <div aria-hidden='true' className={moStyles.overlay} onMouseDown={handleOverlayClick} />
   );
